Refactor #75 - Move `domain.tag.*` -> `data.[model|repository].*`
