+++ app/src/main/java/com/nlab/reminder/domain/common/data/repository/TagRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 The N's lab Open Source Project
+ * Copyright (C) 2023 The N's lab Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,18 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.domain.common.tag
+package com.nlab.reminder.domain.common.data.repository
 
-import android.os.Parcelable
-import androidx.annotation.Keep
-import androidx.compose.runtime.Stable
-import com.nlab.reminder.core.util.test.annotation.ExcludeFromGeneratedTestReport
-import kotlinx.parcelize.Parcelize
+import kotlinx.coroutines.flow.Flow
+import com.nlab.reminder.core.kotlin.util.Result
+import com.nlab.reminder.domain.common.data.model.Tag
 
 /**
  * @author Doohyun
  */
-@Keep   // for navigation graph.
-@Parcelize
-@Stable
-@ExcludeFromGeneratedTestReport
-data class Tag(val tagId: Long, val name: String) : Parcelable
+interface TagRepository {
+    fun get(): Flow<List<Tag>>
+    suspend fun getUsageCount(tag: Tag): Result<Long>
+    suspend fun updateName(tag: Tag, name: String): Result<Unit>
+    suspend fun delete(tag: Tag): Result<Unit>
+}
