+++ app/src/main/java/com/nlab/reminder/domain/common/schedule/ScheduleUiState.kt
@@ -18,7 +18,7 @@ package com.nlab.reminder.domain.common.schedule
 
 import com.nlab.reminder.domain.common.util.link.LinkMetadata
 import com.nlab.reminder.core.util.test.annotation.ExcludeFromGeneratedTestReport
-import com.nlab.reminder.domain.common.tag.Tag
+import com.nlab.reminder.domain.common.data.model.Tag
 
 /**
  * @author Doohyun
