+++ app/src/main/java/com/nlab/reminder/domain/feature/home/HomeAction.kt
@@ -16,7 +16,7 @@
 
 package com.nlab.reminder.domain.feature.home
 
-import com.nlab.reminder.domain.common.tag.Tag
+import com.nlab.reminder.domain.common.data.model.Tag
 import com.nlab.statekit.Action
 
 /**
