+++ app/src/main/java/com/nlab/reminder/domain/feature/home/HomeSideEffect.kt
@@ -17,7 +17,7 @@ package com.nlab.reminder.domain.feature.home
 
 import com.nlab.reminder.core.effect.SideEffect
 import com.nlab.reminder.core.util.test.annotation.ExcludeFromGeneratedTestReport
-import com.nlab.reminder.domain.common.tag.Tag
+import com.nlab.reminder.domain.common.data.model.Tag
 
 /**
  * @author thalys
