+++ app/src/main/java/com/nlab/reminder/domain/feature/home/di/HomeViewModelModule.kt
@@ -19,7 +19,7 @@ package com.nlab.reminder.domain.feature.home.di
 import com.nlab.reminder.core.effect.SideEffectHandle
 import com.nlab.reminder.core.state.StateContainer
 import com.nlab.reminder.core.state.asContainer
-import com.nlab.reminder.domain.common.tag.TagRepository
+import com.nlab.reminder.domain.common.data.repository.TagRepository
 import com.nlab.reminder.domain.feature.home.*
 import dagger.Module
 import dagger.Provides
