+++ app/src/main/java/com/nlab/reminder/internal/common/android/database/TagEntities.kt
@@ -16,7 +16,7 @@
 
 package com.nlab.reminder.internal.common.android.database
 
-import com.nlab.reminder.domain.common.tag.Tag
+import com.nlab.reminder.domain.common.data.model.Tag
 
 /**
  * @author Doohyun
