+++ app/src/main/java/com/nlab/reminder/internal/common/data/repository/LocalTagRepository.kt
@@ -14,20 +14,21 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.internal.common.tag.impl
+package com.nlab.reminder.internal.common.data.repository
 
 import com.nlab.reminder.core.kotlin.coroutine.flow.map
 import com.nlab.reminder.core.kotlin.util.Result
 import com.nlab.reminder.core.kotlin.util.catching
-import com.nlab.reminder.domain.common.tag.Tag
-import com.nlab.reminder.domain.common.tag.TagRepository
+import com.nlab.reminder.domain.common.data.model.Tag
+import com.nlab.reminder.domain.common.data.repository.TagRepository
 import com.nlab.reminder.internal.common.android.database.*
 import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
 
 /**
  * @author Doohyun
  */
-class LocalTagRepository(
+internal class LocalTagRepository @Inject constructor(
     private val tagDao: TagDao,
     private val scheduleTagListDao: ScheduleTagListDao
 ) : TagRepository {
