+++ app/src/main/res/navigation/nav_home.xml
@@ -47,7 +47,7 @@
 			app:nullable="false" />
 		<argument
 			android:name="tag"
-			app:argType="com.nlab.reminder.domain.common.tag.Tag"
+			app:argType="com.nlab.reminder.domain.common.data.model.Tag"
 			app:nullable="false" />
 	</dialog>
 
@@ -62,7 +62,7 @@
 			app:nullable="false" />
 		<argument
 			android:name="tag"
-			app:argType="com.nlab.reminder.domain.common.tag.Tag"
+			app:argType="com.nlab.reminder.domain.common.data.model.Tag"
 			app:nullable="false" />
 		<argument
 			android:name="tagUsageCount"
@@ -81,7 +81,7 @@
 			app:nullable="false" />
 		<argument
 			android:name="tag"
-			app:argType="com.nlab.reminder.domain.common.tag.Tag"
+			app:argType="com.nlab.reminder.domain.common.data.model.Tag"
 			app:nullable="false" />
 		<argument
 			android:name="tagUsageCount"
