+++ app/src/test/java/com/nlab/reminder/domain/common/schedule/ScheduleGenerator.kt
@@ -18,8 +18,8 @@ package com.nlab.reminder.domain.common.schedule
 
 import com.nlab.reminder.domain.common.util.link.LinkMetadata
 import com.nlab.reminder.domain.common.util.transaction.TransactionId
-import com.nlab.reminder.domain.common.tag.Tag
-import com.nlab.reminder.domain.common.tag.genTags
+import com.nlab.reminder.domain.common.data.model.Tag
+import com.nlab.reminder.domain.common.data.model.genTags
 import com.nlab.testkit.*
 
 /**
