+++ app/src/test/java/com/nlab/reminder/domain/feature/home/HomeGenerator.kt
@@ -17,10 +17,10 @@
 package com.nlab.reminder.domain.feature.home
 
 import com.nlab.reminder.core.effect.SideEffectHandle
-import com.nlab.reminder.domain.common.tag.Tag
-import com.nlab.reminder.domain.common.tag.TagRepository
-import com.nlab.reminder.domain.common.tag.genTag
-import com.nlab.reminder.domain.common.tag.genTags
+import com.nlab.reminder.domain.common.data.model.Tag
+import com.nlab.reminder.domain.common.data.repository.TagRepository
+import com.nlab.reminder.domain.common.data.model.genTag
+import com.nlab.reminder.domain.common.data.model.genTags
 import com.nlab.testkit.genBothify
 import com.nlab.testkit.genLong
 import kotlinx.coroutines.flow.emptyFlow
